Lesson 3: Header View – SwiftUI To Do List
In this video we will learn how to make a SwiftUI view generic and reusable. Specifically we will work with the header view that will be common across the Log In and Register screens. We'll leverage parameters to make this component configurable and flexible. By the end of this video, you'll have mastered proper view reusability in SwiftUI.