Lesson 4: Register View – SwiftUI To Do List
In this video we will continue building the New User Registration view / screen. We'll add a form, button, and tweak the header for this view. By the end of this video you will have the logged out state of your app complete. You'll also have a NavigationView working to transition between different views / screens.