Lesson 8: New Item View – SwiftUI To Do List
In this video we will create the New To Do List Item View. This view will enable users to create new items with a title and due date. Moreover, we will add a beautiful calendar date picker that enables users to select a due date. Lastly, we'll discuss information validation to ensure the user is saving a properly made entry.