Lesson 9: Saving To Do Items – SwiftUI To Do List
In this video we will learn to save new To Do List items to the database. Moreover, we'll discuss input validation for both the new item title and due date selection. Lastly, by the end of this video, your newly created to do list items we'll start to appear in the database as you create them.