Lesson 10: Items List – SwiftUI To Do List
In this video we will build our to do list items view. This view will show us a list of our to do list items for the current user. Moreover, this list will update in real-time with a Firestore query listener. In addition, we will showcase the due date for each item in the list – as well as a checkmark circle too mark items as complete/incomplete.