Lesson 11: Delete Items – SwiftUI To Do List
In this video we will enable the Swipe to Delete functionality for a user's created to do list items. This common pattern is used across hundreds of apps for item deletion. We'll hook up the delete action to the database and watch our items vanish in real-time as we delete them.